requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.0
greenlet==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==10.1.0
pypng==0.20220715.0
qrcode==7.4.2
reportlab==4.0.7
SQLAlchemy==2.0.23
typing_extensions==4.8.0
Werkzeug==3.0.1
